+++ CHANGES.rst
@@ -648,6 +648,9 @@ API Changes
     ``ValueError`` only when the ``bbox`` attribute was accessed.
     [#XXXX]
 
+  - ``SegmentationImage`` now raises a ``TypeError`` for masked array
+    input. [#2381]
+
 - ``photutils.utils``
 
   - The ``ShepardIDWInterpolator`` ``ncoords`` attribute has been